requirements.txt
discord.py==2.3.0
flask==2.3.0
aiohttp==3.8.0